third_party/blink/web_tests/external/wpt/css/css-text/i18n/unknown-lang/css-text-line-break-iteration-loose-expected.txt

This is a testharness.js-based test.
[FAIL] 3005  IDEOGRAPHIC ITERATION MARK may appear at line start if lang unknown and loose
  assert_approx_equals: expected 69 +/- 1 but got 99
[FAIL] 303B  VERTICAL IDEOGRAPHIC ITERATION MARK may appear at line start if lang unknown and loose
  assert_approx_equals: expected 69 +/- 1 but got 99
[FAIL] 309D  HIRAGANA ITERATION MARK may appear at line start if lang unknown and loose
  assert_approx_equals: expected 69 +/- 1 but got 99
[FAIL] 309E  HIRAGANA VOICED ITERATION MARK may appear at line start if lang unknown and loose
  assert_approx_equals: expected 69 +/- 1 but got 99
[FAIL] 30FD  KATAKANA ITERATION MARK may appear at line start if lang unknown and loose
  assert_approx_equals: expected 69 +/- 1 but got 99
[FAIL] 30FE  KATAKANA VOICED ITERATION MARK may appear at line start if lang unknown and loose
  assert_approx_equals: expected 69 +/- 1 but got 99
Harness: the test ran to completion.
